+++ react-demo-app/src/features/conversationsSlice.ts
@@ -1,20 +1,29 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { IPendingSession, IStoredConversationState } from '../../../dist/es';
+import {IParticipantsUpdate, VideoMediaSession} from "../../../src";
 
 interface IConversationsState {
   pendingSessions: IPendingSession[],
   handledPendingSessions: IPendingSession[],
-  activeConversations: IActiveConversationsState
+  activeConversations: IActiveConversationsState,
+  activeVideoConversations: IActiveVideoConversationsState[]
 }
 
 interface IActiveConversationsState {
   [key: string]: IStoredConversationState;
 }
 
+export interface IActiveVideoConversationsState {
+    conversationId: string;
+    session: VideoMediaSession;
+    participantsUpdate: IParticipantsUpdate;
+}
+
 const initialState: IConversationsState = {
   pendingSessions: [],
   handledPendingSessions: [],
-  activeConversations: {}
+  activeConversations: {},
+  activeVideoConversations: []
 }
 
 export const conversationsSlice = createSlice({
@@ -46,9 +55,37 @@ export const conversationsSlice = createSlice({
       if (!existingSession) {
         state.handledPendingSessions = [...state.handledPendingSessions, action.payload];
       }
-    }
+    },
+    addVideoConversationToActive: (state, action) => {
+      state.activeVideoConversations = [...state.activeVideoConversations, action.payload]
+    },
+    addParticipantUpdateToVideoConversation: (state, action) => {
+      const conv = state.activeVideoConversations.find(
+        conv => conv.conversationId === action.payload.conversationId);
+      if (conv) {
+        conv.participantsUpdate = action.payload;
+      }
+      state.activeVideoConversations = [...state.activeVideoConversations];
+    },
+    removeVideoConversationFromActive: (state, action) => {
+      const newFilteredState = state.activeVideoConversations.filter(
+        convo => action.payload.conversationId !== convo.conversationId);
+      state.activeVideoConversations = [...newFilteredState];
+    },
+    reasignToTriggerRepaint: (state) => {
+      state.activeVideoConversations = [...state.activeVideoConversations];
+    },
   }
-})
+});
 
-export const { updatePendingSessions, removePendingSession, updateConversations, storeHandledPendingSession } = conversationsSlice.actions;
+export const {
+  updatePendingSessions,
+  removePendingSession,
+  updateConversations,
+  storeHandledPendingSession,
+  addVideoConversationToActive,
+  addParticipantUpdateToVideoConversation,
+  removeVideoConversationFromActive,
+  reasignToTriggerRepaint,
+} = conversationsSlice.actions;
 export default conversationsSlice.reducer;