+++ react-demo-app/src/hooks/useSdk.ts
@@ -55,7 +55,7 @@ export default function useSdk() {
     webrtcSdk.on('pendingSession', handlePendingSession);
     webrtcSdk.on('cancelPendingSession', handleCancelPendingSession);
     webrtcSdk.on('handledPendingSession', handledPendingSession);
-    webrtcSdk.on('sessionStarted', handleSessionStarted);
+    // webrtcSdk.on('sessionStarted', handleSessionStarted);
     webrtcSdk.on('sessionEnded', (session) => handleSessionEnded(session));
     // webrtcSdk.on('trace', trace);
     webrtcSdk.on('disconnected', handleDisconnected);
@@ -106,6 +106,9 @@ export default function useSdk() {
   }
   async function toggleAudioMute(mute: boolean, conversationId: string): Promise<void> {
     await sdk.setAudioMute({ mute, conversationId });
+  }
+  async function toggleVideoMute(mute: boolean, conversationId: string): Promise<void> {
+    await sdk.setVideoMute({mute, conversationId});
   }
   async function toggleHoldState(held: boolean, conversationId: string): Promise<void> {
     await sdk.setConversationHeld({ held, conversationId });
@@ -172,6 +175,7 @@ export default function useSdk() {
     startSoftphoneSession,
     endSession,
     toggleAudioMute,
+    toggleVideoMute,
     toggleHoldState,
     updateDefaultDevices,
     enumerateDevices,